refactor(workbench): extract session action modals into WorkbenchSessionModals
Move ask-user, status/usage, dashboards, rewind/fork/resume, plan history, json schema, and session note/rules/turns/prompt dialogs out of the workbench return. Leave CompactModal, QueueEditModal, SettingsPage, and AppDialogHost in place after the splice.
