+++ docs/plans/2026-08-23-wp-w16-pi-review.md
@@ -1,0 +1,11 @@
+## 结论：**未完成（pi 不可用）**
+
+审查对象：`WorkbenchSessionModals`（AskUser → session sys-prompt discard）。
+
+按 AGENTS.md §8 调用 `pi -p`。本机 PowerShell 报：`Get-Command pi` 空。未用其它模型顶替审查。
+
+**本项不得标为已 pi 审过。**
+
+自测：`pnpm exec tsc -b` 过；`python scripts/check-code-quality-gates.py --mode final` → **PASS**。度量 `App.tsx.lines=19862`（shell 18 + AppWorkbench 19844）、`useState=224`、`useEffect=86`、`files_ge_1000=69`。哨兵：`CompactModal` / `QueueEditModal` / `SettingsPage` / `AppDialogHost` 仍在 AppWorkbench。
+
+未宣称验收完成：改这些弹层排版可不打开 workbench；open/send 协议与 chat 体仍在 composition root。
