+++ src/app/AppWorkbench.tsx
@@ -304,7 +304,7 @@ import {
   formatPermissionSummary,
   mapPermissionButtons
 } from "@/lib/permissionOptions";
-import { AskUserModal, dropAskUserClocks } from "@/components/AskUserModal";
+import { dropAskUserClocks } from "@/components/AskUserModal";
 import { type PaletteActionDef } from "@/lib/paletteActions";
 import {
   canOfferContinueCwd,
@@ -444,7 +444,6 @@ import {
 } from "@/lib/draftDoc";
 import {
   isActiveJsonSchema,
-  parseJsonSchemaText,
   wrapAgentTextWithJsonSchema
 } from "@/lib/jsonSchema";
 import { sanitizeExtraRules } from "@/lib/sessionExtraRules";
@@ -616,7 +615,6 @@ import {
   isWorktreeNameCollisionError,
   resolveForkAgentCheckbox,
   resolveForkAgentSession,
-  resolveForkCliOnConfirm,
   resolveSessionForkSoftFail,
   softFailKindFromRestoreGate
 } from "@/lib/sessionFork";
@@ -666,8 +664,6 @@ import {
   uploadMatchesQuery
 } from "@/components/ComposerPlusPanel";
 import { planInsertSkill } from "@/lib/skillsTaskPicker";
-import { StatusModal } from "@/components/StatusModal";
-import { UsageLimitModal } from "@/components/UsageLimitModal";
 import {
   IconChevronDown,
   IconChevronUp,
@@ -768,21 +764,8 @@ import {
 import { applySideContextOpen } from "@/lib/sideContextOpen";
 import { resolveSidePathDeepLink } from "@/lib/sidePathDeepLink";
 
-import { RewindConfirmModal } from "@/components/workbench-modals/RewindConfirmModal";
-import { ForkConfirmModal } from "@/components/workbench-modals/ForkConfirmModal";
-import { ResumeRestoreConfirmModal } from "@/components/workbench-modals/ResumeRestoreConfirmModal";
 import { ConfirmCopyModal } from "@/components/workbench-modals/ConfirmCopyModal";
-import { RewindTimelineModal } from "@/components/workbench-modals/RewindTimelineModal";
-import { TracesHistoryModal } from "@/components/workbench-modals/TracesHistoryModal";
-import { PlanHistoryModal } from "@/components/workbench-modals/PlanHistoryModal";
-import { PlanHistoryPreviewModal } from "@/components/workbench-modals/PlanHistoryPreviewModal";
-import { PlanReviseModal } from "@/components/workbench-modals/PlanReviseModal";
-import { JsonSchemaModal } from "@/components/workbench-modals/JsonSchemaModal";
 import { QueueEditModal } from "@/components/workbench-modals/QueueEditModal";
-import { SessionNoteModal } from "@/components/workbench-modals/SessionNoteModal";
-import { SessionRulesModal } from "@/components/workbench-modals/SessionRulesModal";
-import { SessionMaxTurnsModal } from "@/components/workbench-modals/SessionMaxTurnsModal";
-import { SessionSysPromptModal } from "@/components/workbench-modals/SessionSysPromptModal";
 import { CompactModal } from "@/components/workbench-modals/CompactModal";
 import { AppDialogHost } from "@/components/workbench-modals/AppDialogHost";
 import {
@@ -807,29 +790,9 @@ const AutomationsPage = lazy(async () => {
   const m = await import("@/components/AutomationsPage");
   return { default: m.AutomationsPage };
 });
-const AgentDashboardModal = lazy(async () => {
-  const m = await import("@/components/AgentDashboardModal");
-  return { default: m.AgentDashboardModal };
-});
-const SessionTaskBoardModal = lazy(async () => {
-  const m = await import("@/components/SessionTaskBoardModal");
-  return { default: m.SessionTaskBoardModal };
-});
-const BatchAgentsModal = lazy(async () => {
-  const m = await import("@/components/BatchAgentsModal");
-  return { default: m.BatchAgentsModal };
-});
-const OpsEntryModal = lazy(async () => {
-  const m = await import("@/components/OpsEntryModal");
-  return { default: m.OpsEntryModal };
-});
 const KanbanBoardPage = lazy(async () => {
   const m = await import("@/components/KanbanBoardPage");
   return { default: m.KanbanBoardPage };
-});
-const McpStatusModal = lazy(async () => {
-  const m = await import("@/components/McpStatusModal");
-  return { default: m.McpStatusModal };
 });
 const SetupWizard = lazy(async () => {
   const m = await import("@/components/SetupWizard");
@@ -938,6 +901,7 @@ import { WorkbenchSidebar } from "@/app/WorkbenchSidebar";
 import { WorkbenchMain } from "@/app/WorkbenchMain";
 import { WorkbenchResourcesAside } from "@/app/WorkbenchResourcesAside";
 import { WorkbenchDomainOverlays } from "@/app/WorkbenchDomainOverlays";
+import { WorkbenchSessionModals } from "@/app/WorkbenchSessionModals";
 import { WorkbenchChromeOverlays } from "@/app/WorkbenchChromeOverlays";
 import { useSessionExportText } from "@/hooks/useSessionExportText";
 import { useSessionExportImage } from "@/hooks/useSessionExportImage";
@@ -18525,542 +18489,152 @@ export function AppWorkbench() {
         }}
       />
 
-      <AskUserModal
-        payload={askUser}
-        timeoutSec={askUserTimeoutSec}
-        labels={{
-          title: tr("askUser.title"),
-          submit: tr("askUser.submit"),
-          cancel: tr("askUser.cancel"),
-          otherPlaceholder: tr("askUser.otherPlaceholder"),
-          freeTextHint: tr("askUser.freeTextHint"),
-          multiHint: tr("askUser.multiHint"),
-          close: tr("common.close"),
-          autoCancelCountdown: tr("askUser.autoCancelCountdown"),
-        }}
-        onSubmit={async (answers) => {
-          if (!askUser) return;
-          try {
-            await api.sessionResolveAskUser({
-              decision: "accepted",
-              answers,
-              rpcId: askUser.rpcId,
-              sessionId: askUser.sessionId,
-            });
-            clearPendingGates(askUser.sessionId);
-            setAskUser(null);
-          } catch (e) {
-            showToast(String(e), 4500);
-          }
-        }}
-        onCancel={async () => {
-          if (!askUser) return;
-          try {
-            await api.sessionResolveAskUser({
-              decision: "cancelled",
-              rpcId: askUser.rpcId,
-              sessionId: askUser.sessionId,
-            });
-          } catch {
-            /* still hide UI */
-          }
-          clearPendingGates(askUser.sessionId);
-          setAskUser(null);
-        }}
-      />
-      <StatusModal
-        open={showStatusModal}
-        locale={locale}
-        sessionId={session.sessionId}
-        agentSessionId={session.agentSessionId}
-        modelId={modelId}
-        effort={effort}
-        mode={mode}
-        policy={policy}
-        projectPath={effectiveProjectPath}
-        messageCount={messages.length}
-        onClose={() => setShowStatusModal(false)}
-      />
-      <UsageLimitModal
-        open={showUsageLimitModal}
-        locale={locale}
-        sessionId={session.sessionId}
-        spend={sessionSpend}
+      <WorkbenchSessionModals
         account={account}
-        customRoute={customRouteActive}
-        turnActive={
-          session.state === "streaming" ||
-          session.state === "awaiting_permission"
-        }
-        onClose={() => setShowUsageLimitModal(false)}
-      />
-      {(agentDashboardOpen) ? (
-      <Suspense fallback={null}>
-      <AgentDashboardModal
-        open={agentDashboardOpen}
+        agentDashboardOpen={agentDashboardOpen}
+        agentDashboardRows={agentDashboardRows}
+        askUser={askUser}
+        askUserTimeoutSec={askUserTimeoutSec}
+        batchAgentsOpen={batchAgentsOpen}
+        clearPendingGates={clearPendingGates}
+        clearSessionMaxTurnsModal={clearSessionMaxTurnsModal}
+        clearSessionRulesModal={clearSessionRulesModal}
+        clearSessionSysPromptModal={clearSessionSysPromptModal}
+        closeSessionMaxTurnsModal={closeSessionMaxTurnsModal}
+        closeSessionNoteModal={closeSessionNoteModal}
+        closeSessionRulesModal={closeSessionRulesModal}
+        closeSessionSysPromptModal={closeSessionSysPromptModal}
+        confirmClearPlanHistory={confirmClearPlanHistory}
+        confirmClearSessionNoteModal={confirmClearSessionNoteModal}
+        confirmRewindToPrompt={confirmRewindToPrompt}
+        customRouteActive={customRouteActive}
+        effectiveProjectPath={effectiveProjectPath}
+        effort={effort}
+        forceCloseSessionNoteModal={forceCloseSessionNoteModal}
+        forceCloseSessionRulesModal={forceCloseSessionRulesModal}
+        forceCloseSessionSysPromptModal={forceCloseSessionSysPromptModal}
+        forkAgentCheckbox={forkAgentCheckbox}
+        forkBusy={forkBusy}
+        forkConfirm={forkConfirm}
+        forkRestoreCode={forkRestoreCode}
+        jsonSchemaDraft={jsonSchemaDraft}
         locale={locale}
-        rows={agentDashboardRows}
-        onClose={() => setAgentDashboardOpen(false)}
-        onSelectSession={(id) => {
-          const row = sessions.find((s) => s.id === id);
-          if (!row) return;
-          const proj = projects.find((p) => p.id === row.projectId) || null;
-          void openSession(row, proj);
-        }}
-        onStopAllBusy={() => stopAllBusySessions("dashboard")}
-        onStopSessions={(ids) => {
-          const n = ids.length;
-          stopBusySessionsByIds(ids, {
-            title: tr("dashboard.stopSelectedTitle", { n }),
-            message: tr("dashboard.stopSelectedConfirm", { n: String(n) }),
-            confirmLabel: tr("dashboard.stopSelected", { n }),
-          });
-        }}
-        onOpenBatchAgents={() => {
-          setAgentDashboardOpen(false);
-          openBatchAgents();
-        }}
-        onOpenTaskBoard={() => {
-          setAgentDashboardOpen(false);
-          setTaskBoardOpen(true);
-        }}
-      />
-      </Suspense>
-      ) : null}
-      {(opsEntryOpen) ? (
-      <Suspense fallback={null}>
-      <OpsEntryModal
-        open={opsEntryOpen}
-        locale={locale}
-        counts={opsEntryCounts}
-        onClose={() => setOpsEntryOpen(false)}
-        onSelect={openOpsDestination}
-      />
-      </Suspense>
-      ) : null}
-      {(taskBoardOpen) ? (
-      <Suspense fallback={null}>
-      <SessionTaskBoardModal
-        open={taskBoardOpen}
-        locale={locale}
-        board={sessionTaskBoard}
-        includeArchived={taskBoardIncludeArchived}
-        onIncludeArchivedChange={setTaskBoardIncludeArchived}
-        onClose={() => setTaskBoardOpen(false)}
-        onSelectSession={(id) => {
-          setTaskBoardOpen(false);
-          const row = sessions.find((s) => s.id === id);
-          if (!row) return;
-          const proj = projects.find((p) => p.id === row.projectId) || null;
-          void openSession(row, proj);
-        }}
-      />
-      </Suspense>
-      ) : null}
-      {(batchAgentsOpen) ? (
-      <Suspense fallback={null}>
-      <BatchAgentsModal
-        open={batchAgentsOpen}
-        locale={locale}
-        projects={projects.map(
-          (p): BatchProjectInput => ({
-            id: p.id,
-            name: p.name,
-            path: p.path,
-            trusted: p.trusted,
-            pathOk: p.pathOk,
-            system: p.system,
-          }),
-        )}
-        onClose={() => setBatchAgentsOpen(false)}
-        onDispatch={runBatchAgentsDispatch}
-      />
-      </Suspense>
-      ) : null}
-      {(showMcpModal) ? (
-      <Suspense fallback={null}>
-      <McpStatusModal
-        open={showMcpModal}
-        locale={locale}
-        servers={mcpServers}
-        error={mcpError}
-        loading={mcpLoading}
-        onClose={() => setShowMcpModal(false)}
-        onManage={() => navigateSettings("extensions")}
-        onRefresh={() => void refreshMcpModal()}
-        doctorReport={mcpDoctorReport}
-        doctorError={mcpDoctorError}
-        doctorLoading={mcpDoctorLoading}
-        doctorFocus={mcpDoctorFocus}
-        onRunDoctor={(name) => void runMcpDoctor(name)}
-        onRefreshDoctor={(name) => runMcpDoctor(name)}
-      />
-      </Suspense>
-      ) : null}
-      <RewindTimelineModal
-        locale={locale}
-        timeline={rewindTimeline}
-        busy={rewindBusy}
-        panelRef={rewindModalRef}
-        onClose={() => setRewindTimeline(null)}
-        onPick={(promptIndex, preview) => {
-          if (!rewindTimeline) return;
-          confirmRewindToPrompt(
-            rewindTimeline.sessionId,
-            promptIndex,
-            preview,
-          );
-        }}
-      />
-
-      <RewindConfirmModal
-        locale={locale}
-        confirm={rewindConfirm}
-        busy={rewindBusy}
-        restoreFiles={rewindRestoreFiles}
-        onRestoreFilesChange={setRewindRestoreFiles}
-        onClose={() => {
-          setRewindConfirm(null);
-          setRewindRestoreFiles(false);
-        }}
-        onConfirm={() => {
-          if (!rewindConfirm) return;
-          void runRewindToPrompt(
-            rewindConfirm.sessionId,
-            rewindConfirm.targetPromptIndex,
-            rewindRestoreFiles,
-          );
-        }}
-      />
-
-      <ForkConfirmModal
-        locale={locale}
-        confirm={forkConfirm}
-        busy={forkBusy}
-        restoreCode={forkRestoreCode}
-        agentCheckbox={forkAgentCheckbox}
-        onRestoreCodeChange={setForkRestoreCode}
-        onForkCliSessionChange={setForkCliSession}
-        onClose={() => {
-          setForkConfirm(null);
-          setForkRestoreCode(false);
-          setForkCliSession(false);
-        }}
-        onConfirm={() => {
-          if (!forkConfirm) return;
-          void runForkSession(forkConfirm.source, {
-            throughUserPromptIndex:
-              forkConfirm.throughUserPromptIndex ?? null,
-            restoreCode: forkRestoreCode,
-            forkCliSession: resolveForkCliOnConfirm({
-              throughUserPromptIndex:
-                forkConfirm.throughUserPromptIndex ?? null,
-              checkboxChecked: forkAgentCheckbox.checked,
-              agentSessionId: forkConfirm.source.agentSessionId,
-            }),
-          });
-        }}
-      />
-
-      <ResumeRestoreConfirmModal
-        locale={locale}
-        open={!!resumeRestoreConfirm}
-        busy={resumeRestoreBusy}
-        agentCheckbox={resumeAgentCheckbox}
-        onForkCliSessionChange={setResumeForkCliSession}
-        onClose={() => {
-          setResumeRestoreConfirm(null);
-          setResumeForkCliSession(false);
-        }}
-        onConfirm={() => {
-          if (!resumeRestoreConfirm) return;
-          void runResumeWithCodeRestore(resumeRestoreConfirm, {
-            forkCliSession: resumeAgentCheckbox.checked,
-          });
-        }}
-      />
-
-      <TracesHistoryModal
-        locale={locale}
-        open={showTraces}
-        onClose={() => setShowTraces(false)}
-        onError={(msg) => showToast(msg, 4000)}
-      />
-
-      <PlanHistoryModal
-        locale={locale}
-        open={showPlanHistory}
-        existingSessionIds={sessions.map((s) => s.id)}
-        onClose={() => setShowPlanHistory(false)}
-        onOpen={(entry) => setPlanHistoryPreview(entry)}
-        onOpenSession={(entry) => openPlanHistorySession(entry)}
-        onRequestClearAll={confirmClearPlanHistory}
-      />
-
-      <PlanHistoryPreviewModal
-        locale={locale}
-        entry={planHistoryPreview}
-        canOpenSession={
-          !!planHistoryPreview &&
-          sessions.some((s) => s.id === planHistoryPreview.sessionId)
-        }
-        onClose={() => setPlanHistoryPreview(null)}
-        onOpenSession={() => {
-          if (planHistoryPreview) openPlanHistorySession(planHistoryPreview);
-        }}
-      />
-
-      <PlanReviseModal
-        locale={locale}
-        open={planReviseOpen}
-        note={planReviseNote}
-        onNoteChange={setPlanReviseNote}
-        onClose={() => {
-          setPlanReviseOpen(false);
-          setPlanReviseNote("");
-        }}
-        onSubmit={() => void requestPlanChanges(planReviseNote)}
-      />
-
-      <JsonSchemaModal
-        locale={locale}
-        open={showJsonSchemaModal}
-        draft={jsonSchemaDraft}
-        hasStoredSchema={!!sessionJsonSchema}
-        onDraftChange={setJsonSchemaDraft}
-        onClose={() => setShowJsonSchemaModal(false)}
-        onClear={() => {
-          void (async () => {
-            const sid = session.sessionId;
-            setSessionJsonSchema(null);
-            setJsonSchemaDraft("");
-            setShowJsonSchemaModal(false);
-            if (sid && api.isTauri()) {
-              try {
-                await api.sessionSetJsonSchema(sid, null);
-              } catch {
-                /* ignore */
-              }
-            }
-            if (sid) {
-              setSessions((list) =>
-                list.map((row) =>
-                  row.id === sid ? { ...row, jsonSchema: null } : row,
-                ),
-              );
-            }
-          })();
-        }}
-        onApply={() => {
-          void (async () => {
-            const parsed = parseJsonSchemaText(jsonSchemaDraft);
-            if (!parsed.ok) {
-              showToast(tr("composer.jsonSchemaInvalid"), 4000);
-              return;
-            }
-            const sid = session.sessionId;
-            setSessionJsonSchema(parsed.normalized);
-            if (sid && api.isTauri()) {
-              try {
-                const saved = await api.sessionSetJsonSchema(
-                  sid,
-                  parsed.normalized,
-                );
-                const next =
-                  typeof saved.jsonSchema === "string" &&
-                  saved.jsonSchema.trim()
-                    ? saved.jsonSchema
-                    : parsed.normalized;
-                setSessionJsonSchema(next);
-                setSessions((list) =>
-                  list.map((row) =>
-                    row.id === sid ? { ...row, jsonSchema: next } : row,
-                  ),
-                );
-              } catch (e) {
-                showToast(String(e), 4000);
-                return;
-              }
-            } else if (!sid) {
-              showToast(tr("composer.jsonSchemaEmptySession"), 3200);
-            }
-            setShowJsonSchemaModal(false);
-          })();
-        }}
-      />
-
-      <SessionNoteModal
-        locale={locale}
-        open={!!sessionNoteTarget}
-        sessionTitle={sessionNoteTarget?.title ?? null}
-        draft={sessionNoteDraft}
-        baseline={sessionNoteBaseline}
-        hadStored={Boolean(
-          sessionNoteTarget && sessionNotesMap[sessionNoteTarget.id]?.trim(),
-        )}
-        showClear={Boolean(
-          sessionNoteTarget &&
-            (sessionNotesMap[sessionNoteTarget.id]?.trim() ||
-              sessionNoteDraft.trim()),
-        )}
-        onClose={closeSessionNoteModal}
-        onSave={saveSessionNoteModal}
-        onClear={requestClearSessionNoteModal}
-        onDraftChange={setSessionNoteDraft}
-      />
-
-      <ConfirmCopyModal
-        open={sessionNoteDiscardOpen}
-        title={tr("resources.discardTitle")}
-        body={tr("session.noteDiscardBody")}
-        closeLabel={tr("common.close")}
-        cancelLabel={tr("common.cancel")}
-        confirmLabel={tr("resources.discardConfirm")}
-        onClose={() => setSessionNoteDiscardOpen(false)}
-        onConfirm={() => {
-          setSessionNoteDiscardOpen(false);
-          forceCloseSessionNoteModal();
-        }}
-      />
-
-      <ConfirmCopyModal
-        open={sessionNoteClearOpen}
-        title={tr("session.noteClearTitle")}
-        body={tr("session.noteClearBody")}
-        closeLabel={tr("common.close")}
-        cancelLabel={tr("common.cancel")}
-        confirmLabel={tr("session.noteClearConfirm")}
-        danger
-        onClose={() => setSessionNoteClearOpen(false)}
-        onConfirm={() => {
-          setSessionNoteClearOpen(false);
-          confirmClearSessionNoteModal();
-        }}
-      />
-
-      <SessionRulesModal
-        locale={locale}
-        open={!!sessionRulesTarget}
-        sessionTitle={sessionRulesTarget?.title ?? null}
-        draft={sessionRulesDraft}
-        baseline={sessionRulesBaseline}
-        hadStored={sessions.some(
-          (row) =>
-            sessionRulesTarget &&
-            row.id === sessionRulesTarget.id &&
-            !!sanitizeExtraRules(row.extraRules),
-        )}
-        showClear={Boolean(
-          sessionRulesTarget &&
-            (sessionRulesDraft.trim() ||
-              sessions.some(
-                (row) =>
-                  row.id === sessionRulesTarget.id &&
-                  !!sanitizeExtraRules(row.extraRules),
-              )),
-        )}
-        busy={sessionRulesBusy}
-        error={sessionRulesError}
-        onClose={closeSessionRulesModal}
-        onSave={() => {
-          void saveSessionRulesModal();
-        }}
-        onClear={() => {
-          void clearSessionRulesModal();
-        }}
-        onDraftChange={(value) => {
-          setSessionRulesDraft(value);
-          setSessionRulesError(null);
-        }}
-      />
-
-      <ConfirmCopyModal
-        open={sessionRulesDiscardOpen}
-        title={tr("resources.discardTitle")}
-        body={tr("session.promptDiscardBody")}
-        closeLabel={tr("common.close")}
-        cancelLabel={tr("common.cancel")}
-        confirmLabel={tr("resources.discardConfirm")}
-        onClose={() => setSessionRulesDiscardOpen(false)}
-        onConfirm={() => {
-          setSessionRulesDiscardOpen(false);
-          forceCloseSessionRulesModal();
-        }}
-      />
-
-      <SessionMaxTurnsModal
-        locale={locale}
-        open={!!sessionMaxTurnsTarget}
-        sessionTitle={sessionMaxTurnsTarget?.title ?? null}
-        draft={sessionMaxTurnsDraft}
-        globalTurns={maxAgentTurns}
-        showClear={Boolean(
-          sessionMaxTurnsTarget &&
-            (sessionMaxTurnsDraft.trim() ||
-              sessions.some(
-                (row) =>
-                  row.id === sessionMaxTurnsTarget.id &&
-                  normalizeMaxAgentTurns(row.maxAgentTurns) != null,
-              )),
-        )}
-        onClose={closeSessionMaxTurnsModal}
-        onSave={() => {
-          void saveSessionMaxTurnsModal();
-        }}
-        onClear={() => {
-          void clearSessionMaxTurnsModal();
-        }}
-        onDraftChange={setSessionMaxTurnsDraft}
-      />
-
-      <SessionSysPromptModal
-        locale={locale}
-        open={!!sessionSysPromptTarget}
-        sessionTitle={sessionSysPromptTarget?.title ?? null}
-        draft={sessionSysPromptDraft}
-        baseline={sessionSysPromptBaseline}
-        hadStored={sessions.some(
-          (row) =>
-            sessionSysPromptTarget &&
-            row.id === sessionSysPromptTarget.id &&
-            !!sanitizeSystemPromptOverride(row.systemPromptOverride),
-        )}
-        showClear={Boolean(
-          sessionSysPromptTarget &&
-            (sessionSysPromptDraft.trim() ||
-              sessions.some(
-                (row) =>
-                  row.id === sessionSysPromptTarget.id &&
-                  !!sanitizeSystemPromptOverride(row.systemPromptOverride),
-              )),
-        )}
-        busy={sessionSysPromptBusy}
-        error={sessionSysPromptError}
-        onClose={closeSessionSysPromptModal}
-        onSave={() => {
-          void saveSessionSysPromptModal();
-        }}
-        onClear={() => {
-          void clearSessionSysPromptModal();
-        }}
-        onDraftChange={(value) => {
-          setSessionSysPromptDraft(value);
-          setSessionSysPromptError(null);
-        }}
-      />
-
-      <ConfirmCopyModal
-        open={sessionSysPromptDiscardOpen}
-        title={tr("resources.discardTitle")}
-        body={tr("session.promptDiscardBody")}
-        closeLabel={tr("common.close")}
-        cancelLabel={tr("common.cancel")}
-        confirmLabel={tr("resources.discardConfirm")}
-        onClose={() => setSessionSysPromptDiscardOpen(false)}
-        onConfirm={() => {
-          setSessionSysPromptDiscardOpen(false);
-          forceCloseSessionSysPromptModal();
-        }}
+        maxAgentTurns={maxAgentTurns}
+        mcpDoctorError={mcpDoctorError}
+        mcpDoctorFocus={mcpDoctorFocus}
+        mcpDoctorLoading={mcpDoctorLoading}
+        mcpDoctorReport={mcpDoctorReport}
+        mcpError={mcpError}
+        mcpLoading={mcpLoading}
+        mcpServers={mcpServers}
+        messages={messages}
+        mode={mode}
+        modelId={modelId}
+        navigateSettings={navigateSettings}
+        openBatchAgents={openBatchAgents}
+        openOpsDestination={openOpsDestination}
+        openPlanHistorySession={openPlanHistorySession}
+        openSession={openSession}
+        opsEntryCounts={opsEntryCounts}
+        opsEntryOpen={opsEntryOpen}
+        planHistoryPreview={planHistoryPreview}
+        planReviseNote={planReviseNote}
+        planReviseOpen={planReviseOpen}
+        policy={policy}
+        projects={projects}
+        refreshMcpModal={refreshMcpModal}
+        requestClearSessionNoteModal={requestClearSessionNoteModal}
+        requestPlanChanges={requestPlanChanges}
+        resumeAgentCheckbox={resumeAgentCheckbox}
+        resumeRestoreBusy={resumeRestoreBusy}
+        resumeRestoreConfirm={resumeRestoreConfirm}
+        rewindBusy={rewindBusy}
+        rewindConfirm={rewindConfirm}
+        rewindModalRef={rewindModalRef}
+        rewindRestoreFiles={rewindRestoreFiles}
+        rewindTimeline={rewindTimeline}
+        runBatchAgentsDispatch={runBatchAgentsDispatch}
+        runForkSession={runForkSession}
+        runMcpDoctor={runMcpDoctor}
+        runResumeWithCodeRestore={runResumeWithCodeRestore}
+        runRewindToPrompt={runRewindToPrompt}
+        saveSessionMaxTurnsModal={saveSessionMaxTurnsModal}
+        saveSessionNoteModal={saveSessionNoteModal}
+        saveSessionRulesModal={saveSessionRulesModal}
+        saveSessionSysPromptModal={saveSessionSysPromptModal}
+        session={session}
+        sessionJsonSchema={sessionJsonSchema}
+        sessionMaxTurnsDraft={sessionMaxTurnsDraft}
+        sessionMaxTurnsTarget={sessionMaxTurnsTarget}
+        sessionNoteBaseline={sessionNoteBaseline}
+        sessionNoteClearOpen={sessionNoteClearOpen}
+        sessionNoteDiscardOpen={sessionNoteDiscardOpen}
+        sessionNoteDraft={sessionNoteDraft}
+        sessionNoteTarget={sessionNoteTarget}
+        sessionNotesMap={sessionNotesMap}
+        sessionRulesBaseline={sessionRulesBaseline}
+        sessionRulesBusy={sessionRulesBusy}
+        sessionRulesDiscardOpen={sessionRulesDiscardOpen}
+        sessionRulesDraft={sessionRulesDraft}
+        sessionRulesError={sessionRulesError}
+        sessionRulesTarget={sessionRulesTarget}
+        sessionSpend={sessionSpend}
+        sessionSysPromptBaseline={sessionSysPromptBaseline}
+        sessionSysPromptBusy={sessionSysPromptBusy}
+        sessionSysPromptDiscardOpen={sessionSysPromptDiscardOpen}
+        sessionSysPromptDraft={sessionSysPromptDraft}
+        sessionSysPromptError={sessionSysPromptError}
+        sessionSysPromptTarget={sessionSysPromptTarget}
+        sessionTaskBoard={sessionTaskBoard}
+        sessions={sessions}
+        setAgentDashboardOpen={setAgentDashboardOpen}
+        setAskUser={setAskUser}
+        setBatchAgentsOpen={setBatchAgentsOpen}
+        setForkCliSession={setForkCliSession}
+        setForkConfirm={setForkConfirm}
+        setForkRestoreCode={setForkRestoreCode}
+        setJsonSchemaDraft={setJsonSchemaDraft}
+        setOpsEntryOpen={setOpsEntryOpen}
+        setPlanHistoryPreview={setPlanHistoryPreview}
+        setPlanReviseNote={setPlanReviseNote}
+        setPlanReviseOpen={setPlanReviseOpen}
+        setResumeForkCliSession={setResumeForkCliSession}
+        setResumeRestoreConfirm={setResumeRestoreConfirm}
+        setRewindConfirm={setRewindConfirm}
+        setRewindRestoreFiles={setRewindRestoreFiles}
+        setRewindTimeline={setRewindTimeline}
+        setSessionJsonSchema={setSessionJsonSchema}
+        setSessionMaxTurnsDraft={setSessionMaxTurnsDraft}
+        setSessionNoteClearOpen={setSessionNoteClearOpen}
+        setSessionNoteDiscardOpen={setSessionNoteDiscardOpen}
+        setSessionNoteDraft={setSessionNoteDraft}
+        setSessionRulesDiscardOpen={setSessionRulesDiscardOpen}
+        setSessionRulesDraft={setSessionRulesDraft}
+        setSessionRulesError={setSessionRulesError}
+        setSessionSysPromptDiscardOpen={setSessionSysPromptDiscardOpen}
+        setSessionSysPromptDraft={setSessionSysPromptDraft}
+        setSessionSysPromptError={setSessionSysPromptError}
+        setSessions={setSessions}
+        setShowJsonSchemaModal={setShowJsonSchemaModal}
+        setShowMcpModal={setShowMcpModal}
+        setShowPlanHistory={setShowPlanHistory}
+        setShowStatusModal={setShowStatusModal}
+        setShowTraces={setShowTraces}
+        setShowUsageLimitModal={setShowUsageLimitModal}
+        setTaskBoardIncludeArchived={setTaskBoardIncludeArchived}
+        setTaskBoardOpen={setTaskBoardOpen}
+        showJsonSchemaModal={showJsonSchemaModal}
+        showMcpModal={showMcpModal}
+        showPlanHistory={showPlanHistory}
+        showStatusModal={showStatusModal}
+        showToast={showToast}
+        showTraces={showTraces}
+        showUsageLimitModal={showUsageLimitModal}
+        stopAllBusySessions={stopAllBusySessions}
+        stopBusySessionsByIds={stopBusySessionsByIds}
+        taskBoardIncludeArchived={taskBoardIncludeArchived}
+        taskBoardOpen={taskBoardOpen}
+        tr={tr}
       />
 
       <WorkbenchDomainOverlays
